requirements.txt
jinja2
click
boto
python-novaclient
pylint
python-daemon
